+++ AllaganLib.Shared/Extensions/StringExtensions.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using System.Text.RegularExpressions;
@@ -10,6 +11,10 @@ public static class StringExtensions
     public static string ToTitleCase(this string npcNameSingular)
     {
         return CultureInfo.CurrentCulture.TextInfo.ToTitleCase(npcNameSingular.ToLower());
+    }
+
+    public static void OpenBrowser(this string url) {
+        Process.Start(new ProcessStartInfo {FileName = url, UseShellExecute = true});
     }
 
     public static string ToParseable(this string input)